Move TestingUtils to root base testing directory (it’s a helper class used only in tests)
