+++ base/src/test/java/uk/ac/ebi/atlas/TestingUtils.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.atlas.utils;
+package uk.ac.ebi.atlas;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
